Making sure our models don't know about price finding internals
