+++ driver/src/withdraw_driver.rs
@@ -85,6 +85,7 @@ mod tests {
             format!("{:x}", state_hash),
             1,
             vec![100; (models::TOKENS * 2) as usize],
+            models::TOKENS,
         );
 
         let contract = SnappContractMock::new();
@@ -155,6 +156,7 @@ mod tests {
             format!("{:x}", state_hash),
             1,
             vec![100; (models::TOKENS * 2) as usize],
+            models::TOKENS,
         );
 
         let db = DbInterfaceMock::new();
@@ -176,6 +178,7 @@ mod tests {
             format!("{:x}", state_hash),
             1,
             vec![100; (models::TOKENS * 2) as usize],
+            models::TOKENS,
         );
 
         let contract = SnappContractMock::new();
@@ -212,6 +215,7 @@ mod tests {
             format!("{:x}", state_hash),
             1,
             vec![100; (models::TOKENS * 2) as usize],
+            models::TOKENS,
         );
         state.decrement_balance(1, 0, 100);
